chore(morpho-sdk): drop simulation sdk dev dependency
