+++ packages/morpho-sdk/src/types/sharedLiquidity.ts
@@ -1,25 +1,5 @@
-import type {
-  BigIntish,
-  Holding,
-  Market,
-  MarketId,
-  MarketParams,
-  Position,
-  Vault,
-  VaultMarketConfig,
-} from "@morpho-org/blue-sdk";
+import type { BigIntish, MarketId, MarketParams } from "@morpho-org/blue-sdk";
 import type { Address } from "viem";
-
-/**
- * Minimal block reference used to fetch reallocation data at a consistent state.
- */
-export interface MinimalBlock {
-  /** Block number used for RPC reads. */
-  readonly number: bigint;
-
-  /** Block timestamp used for market accrual and allocator-delay calculations. */
-  readonly timestamp: bigint;
-}
 
 /**
  * Options controlling public allocator withdrawal discovery.
@@ -74,35 +54,6 @@ export interface PublicReallocation {
 
   /** Assets to withdraw from the source market. */
   readonly assets: bigint;
-}
-
-/**
- * Input state required by {@link ReallocationData} to compute reallocations.
- */
-export interface InputReallocationData {
-  /** Chain id associated with the fetched state. */
-  readonly chainId: number;
-
-  /** Markets indexed by market id. */
-  readonly markets?: Readonly<Record<MarketId, Market | undefined>>;
-
-  /** Vaults indexed by vault address. */
-  readonly vaults?: Readonly<Record<Address, Vault | undefined>>;
-
-  /** Positions indexed by user or vault address, then by market id. */
-  readonly positions?: Readonly<
-    Record<Address, Readonly<Record<MarketId, Position | undefined>>>
-  >;
-
-  /** Holdings indexed by user or vault address, then by token address. */
-  readonly holdings?: Readonly<
-    Record<Address, Readonly<Record<Address, Holding | undefined>>>
-  >;
-
-  /** Vault market configs indexed by vault address, then by market id. */
-  readonly vaultMarketConfigs?: Readonly<
-    Record<Address, Readonly<Record<MarketId, VaultMarketConfig | undefined>>>
-  >;
 }
 
 /** A single withdrawal from a source market within a vault reallocation. */
